+++ tests/test_ordering.py
@@ -1,24 +1,4 @@
 # -*- coding: utf-8 -*-
-import re
-
-import pytest
-
-pytest_plugins = ['pytester']
-
-
-@pytest.fixture
-def item_names_for(testdir):
-
-    def _item_names_for(tests_content):
-        # some strange code to extract sorted items
-        items = testdir.getitems(tests_content)
-        hook = testdir.config.hook
-        hook.pytest_collection_modifyitems(session=items[0].session,
-                                           config=testdir.config, items=items)
-        return [item.name for item in items]
-
-    return _item_names_for
-
 
 def test_no_marks(item_names_for):
     tests_content = """
@@ -28,48 +8,6 @@ def test_no_marks(item_names_for):
     """
 
     assert item_names_for(tests_content) == ['test_1', 'test_2']
-
-
-def test_first_mark(item_names_for):
-    tests_content = """
-    import pytest
-
-    def test_1(): pass
-
-    @pytest.mark.first
-    def test_2(): pass
-    """
-
-    assert item_names_for(tests_content) == ['test_2', 'test_1']
-
-
-def test_last_mark(item_names_for):
-    tests_content = """
-    import pytest
-
-    @pytest.mark.last
-    def test_1(): pass
-
-    def test_2(): pass
-    """
-
-    assert item_names_for(tests_content) == ['test_2', 'test_1']
-
-
-def test_first_last_marks(item_names_for):
-    tests_content = """
-    import pytest
-
-    @pytest.mark.last
-    def test_1(): pass
-
-    @pytest.mark.first
-    def test_2(): pass
-
-    def test_3(): pass
-    """
-
-    assert item_names_for(tests_content) == ['test_2', 'test_3', 'test_1']
 
 
 def test_order_marks(item_names_for):
@@ -266,9 +204,3 @@ def test_order_mark_class(item_names_for):
     """
 
     assert item_names_for(tests_content) == ['test_3', 'test_4', 'test_5', 'test_1', 'test_2']
-
-
-def test_run_marker_registered(capsys):
-    pytest.main('--markers')
-    out, err = capsys.readouterr()
-    assert '@pytest.mark.run' in out
